Refactored model templates to allow renaming fields
