+++ .gitignore
@@ -11,4 +11,6 @@ Cargo.lock
 docs/source/generated
 docs/build
 
+src/model_generated.rs
+
 venv
