+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "py-horned-owl"
-version = "1.1.0"
+version = "1.2.0"
 authors = ["Janna Hastings <j.hastings@ucl.ac.uk>"]
 edition = "2018"
