+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["maturin>=1.2,<2.0"]
 build-backend = "maturin"
 
 [project]
+dynamic = ["version"]
 name = "py-horned-owl"
-version = "1.1.0"
 requires-python = ">=3.7"
 license = {text = "LGPL-3.0-or-later"}
 classifiers = [
