--- template/templates/header.jinja2
@@ -1,8 +1,0 @@
-use std::{borrow::Borrow, collections::{BTreeSet,hash_map::DefaultHasher}, sync::Arc};
-use std::fmt::{Display, Write};
-use std::hash::{Hash, Hasher};
-
-use horned_owl::model::ArcStr;
-use paste::paste;
-use pyo3::{exceptions::PyKeyError, prelude::*, PyObject, types::{IntoPyDict, PyType}};
-use regex::Regex;
