+++ template/templates/named-struct.rs.jinja2
@@ -1,6 +1,6 @@
 #[doc = concat!("{{ model.name }}(
 {%- for k, v in model.fields.items() -%}
-        {{ k }}: {{ v | as_py_type }},
+        {{ k | py_field(v) }}: {{ v | as_py_type }},
 {%- endfor -%})",
     "\n\n",
     doc!({{ model.name }})
@@ -9,8 +9,9 @@
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct {{ model.name }} {
     {%- for field, type in model.fields.items() %}
+        #[doc="{{ field | py_field(type) }}: {{ type | as_py_type }}"]
         #[pyo3(get,set)]
-        pub {{ field }}: {{ type | as_rust_type }},
+        pub {{ field | py_field(type) }}: {{ type | as_rust_type }},
     {% endfor -%}
 }
 
@@ -19,12 +20,12 @@ impl {{ model.name }} {
     #[new]
     fn new(
 {%- for k, v in model.fields.items() -%}
-        {{ k }}: {{ v | as_rust_type }},
+        {{ k | py_field(v) }}: {{ v | as_rust_type }},
 {%- endfor -%}
 ) -> Self {
         {{ model.name }} {
-            {%- for k in model.fields.keys() %}
-                {{ k }},
+            {%- for k, v in model.fields.items() %}
+                {{ k | py_field(v) }},
             {%- endfor %}
         }
     }
@@ -32,7 +33,7 @@ impl {{ model.name }} {
     fn __getitem__(&self, py: Python<'_>, name: &str) -> PyResult<PyObject> {
         match name {
         {%- for field,type in model.fields.items() %}
-            "{{ field }}" => Ok(self.{{ field }}.clone().into_py(py)),
+            "{{ field | py_field(type) }}" => Ok(self.{{ field | py_field(type) }}.clone().into_py(py)),
         {%- endfor %}
             &_ => Err(PyKeyError::new_err(format!("The field '{}' does not exist.", name)))
         }
@@ -41,8 +42,8 @@ impl {{ model.name }} {
     fn __setitem__(&mut self, name: &str, value: &Bound<'_, PyAny>) -> PyResult<()> {
         match name {
         {%- for field,type in model.fields.items() %}
-            "{{ field }}" => {
-                self.{{ field }} = value.extract()?;
+            "{{ field | py_field(type) }}" => {
+                self.{{ field | py_field(type) }} = value.extract()?;
                 Ok(())
             },
         {%- endfor %}
@@ -86,7 +87,7 @@ impl From<&horned_owl::model::{{ model.name }}<ArcStr>> for {{ model.name }} {
     fn from(value: &horned_owl::model::{{ model.name }}<ArcStr>) -> Self {
         {{ model.name }} {
         {%- for field, type in model.fields.items() %}
-            {{ field }}: IntoCompatible::<{{ type | as_rust_type }}>::into_c(value.{{ field }}.borrow()),
+            {{ field | py_field(type) }}: IntoCompatible::<{{ type | as_rust_type }}>::into_c(value.{{ field }}.borrow()),
         {%- endfor %}
         }
     }
@@ -97,10 +98,10 @@ impl From<&{{ model.name }}> for horned_owl::model::{{ model.name }}<ArcStr> {
     fn from(value: &{{ model.name }}) -> Self {
         horned_owl::model::{{ model.name }}::<ArcStr> {
         {%- for field, type in model.fields.items() %}
-            {{ field }}: value.{{ field }}.borrow().into_c(),
+            {{ field }}: value.{{ field  | py_field(type) }}.borrow().into_c(),
         {%- endfor %}
         }
     }
 }
 
-{% include 'base.jinja2' %}
+{% include 'base.rs.jinja2' %}
