+++ template/templates/static.pyi
@@ -1,0 +1,19 @@
+import typing
+from typing import *
+
+class IRI:
+    parse: Any
+    ...
+
+class Facet:
+    Length: Facet
+    MinLength: Facet
+    MaxLength: Facet
+    Pattern: Facet
+    MinInclusive: Facet
+    MinExclusive: Facet
+    MaxInclusive: Facet
+    MaxExclusive: Facet
+    TotalDigits: Facet
+    FractionDigits: Facet
+    LangRange: Facet
